requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
docopt==0.6.2
idna==2.8
Js2Py==0.67
packaging==20.1
Pillow==7.0.0
pipwin==0.4.6
PyAudio==0.2.11
pyjsparser==2.7.1
pyparsing==2.4.6
PyPrind==2.11.2
pytz==2019.3
requests==2.22.0
robobrowser==0.5.3
six==1.14.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==0.16.1
